requirements.txt
altgraph==0.17.4
bcrypt==4.2.0
cffi==1.16.0
cryptography==43.0.0
customtkinter==5.2.2
darkdetect==0.8.0
librouteros==3.2.1
packaging==24.1
paramiko==3.4.0
pillow==10.4.0
plyer==2.1.0
pycparser==2.22
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyNaCl==1.5.0
schedule==1.2.2
setuptools==71.1.0
